src/server/requirements.txt
fastapi==0.95.0
gunicorn==20.0.4
nltk==3.8.1
passlib[bcrypt]==1.7.4
pydantic==1.8.2
python-jose[cryptography]==3.3.0
python-multipart
sqlalchemy==2.0.16
uvicorn==0.21.1